+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ## 1.4 branch
 ### 1.4.4
 * added `Filesystem` class, all filesystem global function are now deprecated;
+* `TestTrait::assertFilePerms()` is deprecated. Use instead
+    `assertFileIsReadable()`/`assertFileIsWritable()`/
+    `assertDirectoryIsReadable()`/`assertDirectoryIsWritable()`;
 * some functions have been moved to `array_functions.php` file.
 
 ### 1.4.3